+++ client/fingerprint/host.go
@@ -1,11 +1,8 @@
 package fingerprint
 
 import (
-	"fmt"
 	"log"
-	"os/exec"
 	"runtime"
-	"strings"
 
 	"github.com/hashicorp/nomad/client/config"
 	"github.com/hashicorp/nomad/nomad/structs"
@@ -38,11 +35,7 @@ func (f *HostFingerprint) Fingerprint(cfg *config.Config, node *structs.Node) (b
 	node.Attributes["kernel.version"] = ""
 
 	if runtime.GOOS != "windows" {
-		out, err := exec.Command("uname", "-r").Output()
-		if err != nil {
-			return false, fmt.Errorf("Failed to run uname: %s", err)
-		}
-		node.Attributes["kernel.version"] = strings.Trim(string(out), "\n")
+		node.Attributes["kernel.version"] = hostInfo.KernelVersion
 	}
 
 	node.Attributes["unique.hostname"] = hostInfo.Hostname